Separate source code from tests
