+++ src/battle/AttackGlobals.py
