+++ src/core/Factory.py
@@ -1,9 +1,9 @@
-from battle.Battle import BattleCog, ToonAttack
-from core.Cog import Cog
-from core.Entity import BattleEntity, Entity
-from core.Toon import Toon
-from gags.Gag import Gag, get_gag_min_max_exp
-from gags.GagGlobals import TRACK
+from src.battle.Battle import BattleCog, ToonAttack
+from src.core.Cog import Cog
+from src.core.Entity import BattleEntity, Entity
+from src.core.Toon import Toon
+from src.gags.Gag import Gag, get_gag_min_max_exp
+from src.gags.GagGlobals import TRACK
 
 
 class EntityFactory:
