+++ src/core/Toon.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Optional
 
-from core.Entity import Entity
-from core.ToonGlobals import DEFAULT_BEAN_COUNT, DEFAULT_BEAN_LIMIT, DEFAULT_HP
-from gags.Gag import Gags
-from gags.GagGlobals import DEFAULT_GAG_LIMIT
+from src.core.Entity import Entity
+from src.core.ToonGlobals import DEFAULT_BEAN_COUNT, DEFAULT_BEAN_LIMIT, DEFAULT_HP
+from src.gags.Gag import Gags
+from src.gags.GagGlobals import DEFAULT_GAG_LIMIT
 
 
 @dataclass
