+++ src/core/ToonGlobals.py
