+++ src/gags/GagGlobals.py
