+++ tests/entity/test_battle_entity_creation.py
@@ -1,7 +1,7 @@
 import pytest
 
-from core.Entity import BattleEntity, Entity
-from core.Factory import BattleEntityFactory, EntityFactory
+from src.core.Entity import BattleEntity, Entity
+from src.core.Factory import BattleEntityFactory, EntityFactory
 
 # Allow pytest to instantiate BattleEntity by "removing" the class's abstract methods
 BattleEntity.__abstractmethods__ = None
