+++ tests/fixtures/battle_fixtures.py
@@ -1,10 +1,10 @@
 import pytest
 
-from battle.AttackGlobals import MULTIPLIER
-from battle.Battle import Battle, BattleToon, RewardCalculator, ToonAttack
-from core.Cog import Cog, get_random_cog
-from core.Factory import BattleCogFactory
-from core.Toon import Toon
+from src.battle.AttackGlobals import MULTIPLIER
+from src.battle.Battle import Battle, BattleToon, RewardCalculator, ToonAttack
+from src.core.Cog import Cog, get_random_cog
+from src.core.Factory import BattleCogFactory
+from src.core.Toon import Toon
 
 
 @pytest.fixture
