+++ tests/gag/test_gag_creation.py
@@ -1,11 +1,11 @@
 
-from battle.AttackGlobals import GROUP
-from core.ToonGlobals import (ASTRO_EXPECTED_AVAILABLE_GAGS,
+from src.battle.AttackGlobals import GROUP
+from src.core.ToonGlobals import (ASTRO_EXPECTED_AVAILABLE_GAGS,
                               ASTRO_EXPECTED_TRACK_LEVELS,
                               ASTRO_EXPECTED_UNLOCKED_GAGS, ASTRO_GAG_COUNT,
                               ASTRO_TRACK_EXPS)
-from gags.Gag import Gag, Gags
-from gags.GagGlobals import (DEFAULT_GAG_COUNT, DEFAULT_TRACK_EXPS_CURRENT,
+from src.gags.Gag import Gag, Gags
+from src.gags.GagGlobals import (DEFAULT_GAG_COUNT, DEFAULT_TRACK_EXPS_CURRENT,
                              DEFAULT_TRACK_LEVELS, GAG, MULTI_TARGET_GAGS,
                              TRACK)
